+++ NEWS.md
@@ -1,7 +1,9 @@
 # FSA 0.6.16 ongoing
+* `extraSS()`: Modified.  Added `sim_names=` and `com_name=` so that simple descriptive names could be given to the model and printed in the heading of the output.  Added checks for whether the complex model appears more complex or not.  Added tests for warning and error messages.
 * `fishR()`: Modfiied.  Fixed bug with `where="news"`.  Added tests.
 * `fitPlot()`: Modified.  Added ability to modify y-axis limits for the nonlinear regression model.  Thanks to Gabriela N. for asking for this.
 * `hoCoef()`: Modified.  Changed `lmobj=` to `object=`, added degrees-of-freedom to the output matrix, streamlined the code, added some checks, and added some tests.
+* `lrt()`: Modified.  Added `sim_names=` and `com_name=` so that simple descriptive names could be given to the model and printed in the heading of the output.  Added checks for whether the complex model appears more complex or not.  Added tests for warning and error messages.
 
 # FSA 0.6.15 Jun15
 * `addZeroCatch()`: Modified.  Deleted extraneous `print()` statement.